reorganize triggers files
