+++ front/temporal/triggers/workflows.ts
@@ -1,2 +1,33 @@
-export * from "./common/workflows";
-export * from "./wakeup/workflows";
+import type { ContentFragmentInputWithFileIdType } from "@app/types/api/internal/assistant";
+import { proxyActivities } from "@temporalio/workflow";
+
+import type * as activities from "./activities";
+
+const { runTriggeredAgentsActivity } = proxyActivities<typeof activities>({
+  startToCloseTimeout: "5 minutes",
+  retry: {
+    nonRetryableErrorTypes: ["TriggerNonRetryableError"],
+  },
+});
+
+export async function agentTriggerWorkflow({
+  userId,
+  workspaceId,
+  triggerId,
+  contentFragment,
+  webhookRequestId,
+}: {
+  userId: string;
+  workspaceId: string;
+  triggerId: string;
+  contentFragment?: ContentFragmentInputWithFileIdType;
+  webhookRequestId?: number;
+}) {
+  await runTriggeredAgentsActivity({
+    userId,
+    workspaceId,
+    triggerId,
+    contentFragment,
+    webhookRequestId,
+  });
+}
