+++ front/temporal/triggers_garbage_collect/client.ts
@@ -1,7 +1,7 @@
 import { getTemporalClientForAgentNamespace } from "@app/lib/temporal";
 import logger from "@app/logger/logger";
-import { QUEUE_NAME } from "@app/temporal/triggers/webhook/config";
-import { webhookCleanupWorkflow } from "@app/temporal/triggers/webhook/workflows";
+import { QUEUE_NAME } from "@app/temporal/triggers_garbage_collect/config";
+import { webhookCleanupWorkflow } from "@app/temporal/triggers_garbage_collect/workflows";
 import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
 import { normalizeError } from "@app/types/shared/utils/error_utils";
