+++ front/temporal/triggers_garbage_collect/config.ts
