+++ front/temporal/triggers_garbage_collect/workflows.ts
@@ -1,4 +1,4 @@
-import type * as activities from "@app/temporal/triggers/webhook/activities";
+import type * as activities from "@app/temporal/triggers_garbage_collect/activities";
 import { proxyActivities } from "@temporalio/workflow";
 
 const { webhookCleanupActivity } = proxyActivities<typeof activities>({
